+++ src/basis.jl
@@ -305,7 +305,7 @@ end
     Return minimum of numbers greater than zero.
 
 # Arguments
-- `arr::Array{Float64, 1}`: A function that takes a one dimensional array as argument.
+- `arr::Vector{Float64}`: A function that takes a one dimensional array as argument.
 
 # Example
 ```julia-repl